requirements.txt
blinker==1.9.0
click==8.1.8
cobble==0.1.4
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
mammoth==1.9.0
MarkupSafe==3.0.2
numpy==2.2.4
pillow==11.2.1
Werkzeug==3.1.3
